fix: enum accepts an anonymous role literal as a does argument
`does` accepts an anonymous role LITERAL, not just a role name.
`enum Level <Off Fatal> does role { ... }` (Lumberjack's shape) took
`role` as the name of a role to look up, found none, and died with
"Unknown role: role" -- blocking any distribution using this shape from
loading at all.

Measured against rakudo: the two clause orderings behave genuinely
differently there. `does Role` BEFORE the value list is a real
declarator trait that composes the role (`.^does` True), but an
anonymous role literal is invalid even in rakudo there ("Invalid
typename 'role'") -- only a role name is accepted in that slot. `does
Role` AFTER the value list (the reported shape) is not special enum
grammar at all: `enum E <a b>` parses as an ordinary term, and the
trailing `does Role` is the general infix `does` operator applied to it
and then sunk -- it never composes anything (`.^does` False,
`.^roles(:local)` empty), and a second trailing `does` even errors as a
non-associative operator.

The fix mirrors both: the before-the-values does-loop is unchanged
(still rejects a role literal, matching rakudo's rejection via a
different error), and a new skip_trailing_enum_does_clause consumes but
discards a single does clause after the value list, so parsing no
longer crashes or fragments into a stray "Unknown role: role" statement
applied to the wrong scope, without pretending to compose something
rakudo itself does not.

Split enum_decl.rs into enum_decl.rs + enum_decl_traits.rs to stay under
the file-size convention.

Closes #8216
