+++ src/parser/primary/misc.rs
@@ -15,12 +15,12 @@ mod reduction;
 
 pub(crate) use anon_decl::next_anon_role_name;
 pub(super) use anon_decl::{
-    anon_class_expr, anon_grammar_expr, anon_role_expr, indirect_method_call,
-    mark_anon_package_decl,
+    anon_class_expr, anon_grammar_expr, indirect_method_call, mark_anon_package_decl,
 };
 pub(super) use lambda::{arrow_lambda, capture_literal};
 pub(super) use reduction::reduction_op;
 
+pub(in crate::parser) use anon_decl::anon_role_expr;
 pub(in crate::parser) use colonpair::{colonpair_expr, wrap_colonpair_sink_source};
 pub(in crate::parser) use lambda::{
     block_or_hash_expr, double_closure_error, parse_block_body, parse_block_body_routine,
